enable MEMORY64 wasm64 with emsdk 4.0.23 and C++ exception decoding
Bump Docker image to emscripten/emsdk:4.0.23 which has proper
MEMORY64 support. Pass -sMEMORY64=1 at configure time via
CMAKE_CXX_FLAGS so the toolchain sets sizeof(void*)=8. Upgrade
to C++17 (required by embind in emsdk 4.x), replace deprecated
--bind with -lembind, and cast Eigen::Index returns to int for
wasm64 embind wire type compatibility.

Add native WASM exception handling (-fwasm-exceptions) with
EXPORT_EXCEPTION_HANDLING_HELPERS so JS callers get readable
C++ exception messages instead of raw pointer integers.

+++ wasm/bindings.cpp
@@ -361,7 +361,7 @@ static val makeRealResult(SolverType& solver)
     val result = val::object();
     result.set("converged", solver.info() == CompInfo::Successful);
     result.set("info", compInfoToString(solver.info()));
-    result.set("numIter", solver.num_iterations());
+    result.set("numIter", (int)solver.num_iterations());
     result.set("numOps", g_progress.opsCompleted);
     if (solver.info() == CompInfo::Successful)
     {
@@ -377,7 +377,7 @@ static val makeComplexResult(SolverType& solver)
     val result = val::object();
     result.set("converged", solver.info() == CompInfo::Successful);
     result.set("info", compInfoToString(solver.info()));
-    result.set("numIter", solver.num_iterations());
+    result.set("numIter", (int)solver.num_iterations());
     result.set("numOps", g_progress.opsCompleted);
     if (solver.info() == CompInfo::Successful)
     {
@@ -394,7 +394,7 @@ static val makeRealResultTyped(SolverType& solver)
     val result = val::object();
     result.set("converged", solver.info() == CompInfo::Successful);
     result.set("info", compInfoToString(solver.info()));
-    result.set("numIter", solver.num_iterations());
+    result.set("numIter", (int)solver.num_iterations());
     result.set("numOps", g_progress.opsCompleted);
     if (solver.info() == CompInfo::Successful)
     {
@@ -410,7 +410,7 @@ static val makeComplexResultTyped(SolverType& solver)
     val result = val::object();
     result.set("converged", solver.info() == CompInfo::Successful);
     result.set("info", compInfoToString(solver.info()));
-    result.set("numIter", solver.num_iterations());
+    result.set("numIter", (int)solver.num_iterations());
     result.set("numOps", g_progress.opsCompleted);
     if (solver.info() == CompInfo::Successful)
     {
@@ -620,7 +620,7 @@ static val davidsonSymEigs(const val& matData, int n, int nev,
     val result = val::object();
     result.set("converged", solver.info() == CompInfo::Successful);
     result.set("info", compInfoToString(solver.info()));
-    result.set("numIter", solver.num_iterations());
+    result.set("numIter", (int)solver.num_iterations());
     result.set("numOps", g_progress.opsCompleted);
     if (solver.info() == CompInfo::Successful)
     {
